mqtt-server/requirements.txt
fastapi
uvicorn[standard]
websockets
paho-mqtt
msgpack
scipy
Pillow
